+++ src/main/java/duke/ui/TextUi.java
@@ -1,0 +1,4 @@
+package ui;
+
+public class TextUi {
+}
